requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
future==0.18.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.4
hyperframe==5.2.0
idna==2.10
imageio==2.27.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lazy_loader==0.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
numpy==1.24.2
openai==0.27.4
opencv-python-headless==4.7.0.72
packaging==23.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.7
pysrt==1.1.2
pytesseract==0.3.10
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
yarl==1.8.2
